Moving delete dialog to its own composable
